refactor: move pagination hook
